--- python/cinn/common.py
@@ -1,41 +1,0 @@
-# Copyright (c) 2021 CINN Authors. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from .core_api.common import (  # noqa: F401
-    BFloat16,
-    Bool,
-    CINNValue,
-    CINNValuePack,
-    DefaultHostTarget,
-    DefaultNVGPUTarget,
-    DefaultTarget,
-    Float,
-    Float16,
-    Int,
-    RefCount,
-    Shared_CINNValuePack_,
-    String,
-    Target,
-    Type,
-    UInt,
-    Void,
-    _CINNValuePack_,
-    get_target,
-    is_compiled_with_cuda,
-    is_compiled_with_cudnn,
-    make_const,
-    reset_name_id,
-    set_target,
-    type_of,
-)
