--- python/cinn/lang.py
@@ -1,32 +1,0 @@
-# Copyright (c) 2021 CINN Authors. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from .core_api.lang import (  # noqa: F401
-    Buffer,
-    Module,
-    Placeholder,
-    ReturnType,
-    call_extern,
-    call_lowered,
-    compute,
-    create_placeholder,
-    lower,
-    lower_vec,
-    reduce_all,
-    reduce_any,
-    reduce_max,
-    reduce_min,
-    reduce_mul,
-    reduce_sum,
-)
