--- python/cinn/pe.py
@@ -1,81 +1,0 @@
-# Copyright (c) 2021 CINN Authors. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from .core_api.pe import (  # noqa: F401
-    abs,
-    acos,
-    acosh,
-    add,
-    asin,
-    asinh,
-    atan,
-    atan2,
-    atanh,
-    bitwise_and,
-    bitwise_not,
-    bitwise_or,
-    bitwise_xor,
-    ceil,
-    cos,
-    cosh,
-    divide,
-    equal,
-    erf,
-    exp,
-    floor,
-    floor_divide,
-    greater,
-    greater_equal,
-    identity,
-    isfinite,
-    isinf,
-    isnan,
-    left_shift,
-    less,
-    less_equal,
-    log,
-    log2,
-    log10,
-    logical_and,
-    logical_not,
-    logical_or,
-    logical_xor,
-    matmul,
-    matmul_mkl,
-    max,
-    min,
-    mod,
-    multiply,
-    negative,
-    not_equal,
-    reduce_all,
-    reduce_any,
-    reduce_max,
-    reduce_min,
-    reduce_prod,
-    reduce_sum,
-    remainder,
-    right_shift,
-    round,
-    rsqrt,
-    sigmoid,
-    sign,
-    sin,
-    sinh,
-    sqrt,
-    subtract,
-    tan,
-    tanh,
-    trunc,
-)
