--- python/cinn/runtime/utils.py
@@ -1,35 +1,0 @@
-# Copyright (c) 2023 PaddlePaddle Authors. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import inspect
-
-
-def get_func_global_vars(func):
-    if inspect.ismethod(func):
-        func = func.__func__
-
-    code = func.__code__
-    global_vars = {}
-    if func.__closure__ is not None:
-        for k, v in zip(code.co_freevars, func.__closure__):
-            global_vars[k] = v.cell_contents
-    return global_vars
-
-
-def inspect_function_scope(func):
-    scope = {
-        **func.__globals__,
-        **get_func_global_vars(func),
-    }
-    return scope
